Backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pydantic==1.10.7